requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
cryptography==45.0.5
cymem==2.0.11
datasets==4.0.0
dill==0.3.8
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.3.0
hf-xet==1.1.5
huggingface-hub==0.34.3
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
langcodes==3.5.0
language_data==1.3.0
lightning-utilities==0.15.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
murmurhash==1.0.13
networkx==3.5
nltk==3.9.1
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
preshed==3.0.10
propcache==0.3.2
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx
pytorch-lightning==2.5.2
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
safetensors==0.5.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.14.0
thinc==8.3.6
tokenizers==0.21.4
torch==2.7.1
torchmetrics==1.8.0
tqdm==4.67.1
transformers==4.54.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
